deploy: support enforcing HTTPS in base URL
The official `git-scm.com` website is currently built by Hugo with
`--baseURL http://git-scm.com/`. This commit allows changing the URL
scheme to `https`, in order to avoid unnecessary redirects between
HTTPS and unencrypted HTTP.

Such redirects occur when one of the following URLs is requested:

* URLs in `layouts/alias.html`, e.g. in <https://git-scm.com/docs/>
* Image URLs in <https://git-scm.com/application.min.css>
* Endpoint URLs in <https://git-scm.com/sitemap.xml>

Since these URLs have the `http` scheme, the client is supposed to send
the new request unencrypted, only to be told by the server to use HTTPS
again. Modern web browsers tend to stick with HTTPS in certain cases,
so the downgrade may not always happen, but it's better to eliminate
the redirects nevertheless, for security and performance reasons.

Instead of trying to use relative URLs everywhere, this commit takes a
simpler approach by using the `https` scheme in the base URL. One way
to do this is to enable the "Enforce HTTPS" option in the GitHub Pages
settings, but it's infeasible because the `git-scm.com` domain points to
Cloudflare instead of GitHub.

Therefore, this commit introduces a CI variable, `EXTERNAL_HTTPS`, which
can be set to true if HTTPS is provided by a third party, as in the case
of `git-scm.com`, so that the URL scheme can be overridden. This works
for the official repository as well as any of its forks.